pyproject.toml
[project]
name = "mlops-assignment2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dvc==3.49.0",
    "fastapi==0.129.2",
    "httpx==0.27.0",
    "kaggle==2.0.0",
    "matplotlib==3.8.4",
    "mlflow==2.16.2",
    "numpy==1.26.4",
    "pathspec==0.11.2",
    "pillow==10.3.0",
    "prometheus-client==0.20.0",
    "pydantic==2.7.1",
    "pytest==8.1.1",
    "pytest-cov==5.0.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "scikit-learn==1.4.2",
    "seaborn==0.13.2",
    "setuptools==82.0.0",
    "structlog==24.1.0",
    "torch==2.10.0",
    "torchvision==0.25.0",
    "tqdm==4.66.2",
    "uvicorn[standard]==0.29.0",
]
